(vt) Moved protobuf-specific code to its own subpackage
- This is to make for easier addition of backends
